+++ common/src/main/java/dev/latvian/kubejs/recipe/RecipeJS.java
@@ -22,9 +22,6 @@ import net.minecraft.world.item.crafting.RecipeSerializer;
 import org.apache.commons.codec.binary.Hex;
 import org.jetbrains.annotations.Nullable;
 
-import java.io.ByteArrayOutputStream;
-import java.io.DataOutputStream;
-import java.io.IOException;
 import java.lang.invoke.MethodHandle;
 import java.math.BigInteger;
 import java.security.MessageDigest;
@@ -35,6 +32,8 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.function.BiFunction;
 import java.util.stream.Collectors;
+
+import static dev.latvian.kubejs.util.UtilsJS.getJsonHashBytes;
 
 /**
  * @author LatvianModder
@@ -365,48 +364,6 @@ public abstract class RecipeJS {
 		return inputItems + " -> " + outputItems;
 	}
 
-	private static void writeJsonHash(DataOutputStream stream, @Nullable JsonElement element) throws IOException {
-		if (element == null || element.isJsonNull()) {
-			stream.writeByte('-');
-		} else if (element instanceof JsonArray) {
-			stream.writeByte('[');
-			for (JsonElement e : (JsonArray) element) {
-				writeJsonHash(stream, e);
-			}
-		} else if (element instanceof JsonObject) {
-			stream.writeByte('{');
-			for (Map.Entry<String, JsonElement> e : ((JsonObject) element).entrySet()) {
-				stream.writeBytes(e.getKey());
-				writeJsonHash(stream, e.getValue());
-			}
-		} else if (element instanceof JsonPrimitive) {
-			stream.writeByte('=');
-			if (((JsonPrimitive) element).isBoolean()) {
-				stream.writeBoolean(element.getAsBoolean());
-			} else if (((JsonPrimitive) element).isNumber()) {
-				stream.writeDouble(element.getAsDouble());
-			} else {
-				stream.writeBytes(element.getAsString());
-			}
-		} else {
-			stream.writeByte('?');
-			stream.writeInt(element.hashCode());
-		}
-	}
-
-	public byte[] getJsonHashBytes() {
-		ByteArrayOutputStream baos = new ByteArrayOutputStream();
-		try {
-			writeJsonHash(new DataOutputStream(baos), json);
-		} catch (IOException ex) {
-			ex.printStackTrace();
-			int h = json.hashCode();
-			return new byte[]{(byte) (h >> 24), (byte) (h >> 16), (byte) (h >> 8), (byte) (h >> 0)};
-		}
-
-		return baos.toByteArray();
-	}
-
 	public String getUniqueId() {
 		if (messageDigest == null) {
 			try {
@@ -417,10 +374,10 @@ public abstract class RecipeJS {
 		}
 
 		if (messageDigest == null) {
-			return new BigInteger(Hex.encodeHexString(getJsonHashBytes()), 16).toString(36);
+			return new BigInteger(Hex.encodeHexString(getJsonHashBytes(json)), 16).toString(36);
 		} else {
 			messageDigest.reset();
-			return new BigInteger(Hex.encodeHexString(messageDigest.digest(getJsonHashBytes())), 16).toString(36);
+			return new BigInteger(Hex.encodeHexString(messageDigest.digest(getJsonHashBytes(json))), 16).toString(36);
 		}
 	}
