+++ common/src/main/java/dev/latvian/kubejs/world/gen/WorldgenAddEventJS.java
@@ -1,9 +1,6 @@
 package dev.latvian.kubejs.world.gen;
 
-import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonPrimitive;
 import com.mojang.serialization.JsonOps;
 import dev.latvian.kubejs.event.StartupEventJS;
 import dev.latvian.kubejs.script.ScriptType;
@@ -13,7 +10,6 @@ import dev.latvian.kubejs.util.UtilsJS;
 import dev.latvian.kubejs.world.gen.filter.BiomeFilter;
 import dev.latvian.kubejs.world.gen.ruletest.AnyRuleTest;
 import dev.latvian.kubejs.world.gen.ruletest.InvertRuleTest;
-import me.shedaniel.architectury.annotations.ExpectPlatform;
 import me.shedaniel.architectury.registry.BiomeModifications;
 import net.minecraft.data.BuiltinRegistries;
 import net.minecraft.resources.ResourceLocation;
@@ -37,18 +33,13 @@ import net.minecraft.world.level.levelgen.structure.templatesystem.BlockStateMat
 import net.minecraft.world.level.levelgen.structure.templatesystem.RuleTest;
 import net.minecraft.world.level.levelgen.structure.templatesystem.TagMatchTest;
 import org.apache.commons.codec.binary.Hex;
-import org.jetbrains.annotations.Nullable;
 
-import java.io.ByteArrayOutputStream;
-import java.io.DataOutputStream;
-import java.io.IOException;
 import java.math.BigInteger;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
-import java.util.Map;
 import java.util.function.Consumer;
 
-// TODO: MAJOR cleanup needed!
+import static dev.latvian.kubejs.util.UtilsJS.getJsonHashBytes;
 
 /**
  * @author LatvianModder
@@ -57,7 +48,9 @@ public class WorldgenAddEventJS extends StartupEventJS {
 
 	private static MessageDigest messageDigest;
 
-	protected void addFeature(BiomeFilter filter, GenerationStep.Decoration decoration, ConfiguredFeature<?, ?> feature) {
+	// TODO: we may need to use some platform weirdness down the line,
+	//  for now i'm content with just registering to BuiltinRegistries directly tho
+	private void addFeature(BiomeFilter filter, GenerationStep.Decoration decoration, ConfiguredFeature<?, ?> feature) {
 		ResourceLocation id = new ResourceLocation("kjs_" + getUniqueId(feature));
 		BuiltinRegistries.register(BuiltinRegistries.CONFIGURED_FEATURE, id, feature);
 		BiomeModifications.postProcessProperties(filter, (ctx, properties) -> {
@@ -65,13 +58,7 @@ public class WorldgenAddEventJS extends StartupEventJS {
 		});
 	}
 
-	// FIXME: we'll probably need to use this weirdness down the line,
-	//  for now i'm content with just registering to BuiltinRegistries directly tho
-	@ExpectPlatform
-	public static void registerFeature0(ResourceLocation id, ConfiguredFeature<?,?> feature) {
-	}
-
-	protected void addEntitySpawn(BiomeFilter filter, MobCategory category, MobSpawnSettings.SpawnerData spawnerData) {
+	private void addEntitySpawn(BiomeFilter filter, MobCategory category, MobSpawnSettings.SpawnerData spawnerData) {
 		BiomeModifications.postProcessProperties(filter, (ctx, properties) -> {
 			properties.getSpawnProperties().addSpawn(category, spawnerData);
 		});
@@ -152,49 +139,7 @@ public class WorldgenAddEventJS extends StartupEventJS {
 		addEntitySpawn(properties._biomes, properties._category, new MobSpawnSettings.SpawnerData(properties._entity, properties.weight, properties.minCount, properties.maxCount));
 	}
 
-	private static void writeJsonHash(DataOutputStream stream, @Nullable JsonElement element) throws IOException {
-		if (element == null || element.isJsonNull()) {
-			stream.writeByte('-');
-		} else if (element instanceof JsonArray) {
-			stream.writeByte('[');
-			for (JsonElement e : (JsonArray) element) {
-				writeJsonHash(stream, e);
-			}
-		} else if (element instanceof JsonObject) {
-			stream.writeByte('{');
-			for (Map.Entry<String, JsonElement> e : ((JsonObject) element).entrySet()) {
-				stream.writeBytes(e.getKey());
-				writeJsonHash(stream, e.getValue());
-			}
-		} else if (element instanceof JsonPrimitive) {
-			stream.writeByte('=');
-			if (((JsonPrimitive) element).isBoolean()) {
-				stream.writeBoolean(element.getAsBoolean());
-			} else if (((JsonPrimitive) element).isNumber()) {
-				stream.writeDouble(element.getAsDouble());
-			} else {
-				stream.writeBytes(element.getAsString());
-			}
-		} else {
-			stream.writeByte('?');
-			stream.writeInt(element.hashCode());
-		}
-	}
-
-	public byte[] getJsonHashBytes(JsonElement json) {
-		ByteArrayOutputStream baos = new ByteArrayOutputStream();
-		try {
-			writeJsonHash(new DataOutputStream(baos), json);
-		} catch (IOException ex) {
-			ex.printStackTrace();
-			int h = json.hashCode();
-			return new byte[]{(byte) (h >> 24), (byte) (h >> 16), (byte) (h >> 8), (byte) (h >> 0)};
-		}
-
-		return baos.toByteArray();
-	}
-
-	public String getUniqueId(ConfiguredFeature<?, ?> feature) {
+	public static String getUniqueId(ConfiguredFeature<?, ?> feature) {
 		if (messageDigest == null) {
 			try {
 				messageDigest = MessageDigest.getInstance("MD5");
